requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
python-telegram-bot>=20.7
apscheduler>=3.10.0
pydantic>=2.5.0
paramiko>=3.4.0
bcrypt>=4.1.0
itsdangerous>=2.1.0
jinja2>=3.1.0
python-dotenv>=1.0.0
